Use pagination query and autokey. Also some refactoring.
Nothing too shocking in this one. Set up a pagination query so we don't load all of the posts at once. Fill it in with req.query.page or leave it blank for page 1. Right now we're ignoring the return of the query, but pagination queries have a lot of extra information we might want to look at.