Playlist model v0.1, supports JSON or Mongoose trickery
Be sure to be careful how you pass the tracks in. You can either
pass an array of serialized JSON objects in the "serializedTracks"
field or you can pass an array of regular objects into the "tracks"
field. In the future we'll decide which one to use. Looking for
feedback from Ken when you make the "Create Playlist" page as to
which one you prefer to use.